refact: change js to tsx
