--- src/components/button/Action.js
@@ -1,17 +1,0 @@
-import react from 'react';
-import {
-    Text,
-    TouchableHighlight,
-} from 'react-native';
-
-import styles from './Button.style';
-
-export default ({onClick, label}) => {
-    return (
-        <TouchableHighlight onPress={onClick}>
-            <Text style={styles.action}>
-                {label}
-            </Text>
-        </TouchableHighlight>
-    )
-};
