--- src/components/display/Display.style.js
@@ -1,23 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-    display: {
-        flex: 1,
-        padding: 20,
-        justifyContent: 'center',
-        alignItems: 'center',
-        width: '60%',
-    },
-    displayValue: {
-        fontSize: 24,
-        color: '#fff',
-        textAlign: 'center',
-    },
-    buttons: {
-        padding: 10,
-        flexDirection: 'row',
-        justifyContent: 'center',
-        width: '100%',
-        alignItems: 'center',
-    }
-});
