--- src/modals/Modal.style.js
@@ -1,14 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-    frame: {
-        flex: 1,
-        alignItems: 'center',
-        justifyContent: 'center',
-        backgroundColor: 'rgba(0,0,0,0.8)',
-    },
-    buttons: {
-        flexDirection: 'row',
-        flexWrap: 'wrap',
-    },
-});
